component separation, electors and green icon
