--- src/pages/desk-list/types.ts
@@ -1,9 +1,0 @@
-import { TypographyOwnProps } from '@mui/material';
-
-export interface IDeskItemLabel {
-  typoProps?: TypographyOwnProps;
-  className?: string;
-  deskValue: string;
-  label: string;
-  statusStyle?: {};
-}
